Better .blend file handling

+++ blendersynth/run/pre_ops.py
@@ -3,7 +3,7 @@ import bpy
 from ..blender import render
 
 
-def on_script_open():
+def on_script_open(delete_existing=True):
     """
     On script open:
 
@@ -13,8 +13,9 @@ def on_script_open():
     """
     render.set_engine("CYCLES")
 
-    bpy.data.objects["Cube"].select_set(True)
-    bpy.data.objects["Light"].select_set(True)
-    bpy.ops.object.delete()
+    if delete_existing:
+        bpy.data.objects["Cube"].select_set(True)
+        bpy.data.objects["Light"].select_set(True)
+        bpy.ops.object.delete()
 
     bpy.context.preferences.view.show_splash = False  # disable splash screen
